feat: move write blocks for server
